+++ src/services/firebase/smp-firebase-service-config.interface.ts
@@ -1,7 +1,7 @@
 import {SmpLoggerMethods} from "@tonysamperi/logger";
 //
-import {SmpSocialSignInKeys} from "../../shared/common/smp-social-sign-in-keys.enum";
-import {SmpSocialSignInUxModes} from "../../shared/common/smp-social-sign-in-ux-modes.enum";
+import {SmpSocialSignInKeys} from "../../shared/common/smp-social-sign-in-keys.enum.js";
+import {SmpSocialSignInUxModes} from "../../shared/common/smp-social-sign-in-ux-modes.enum.js";
 
 export interface SmpFirebaseServiceConfig {
     enabledSocials?: SmpSocialSignInKeys[];