+++ src/shared/crypto/index.ts
@@ -1,3 +1,4 @@
 export {SmpAesCrypt} from "./smp-aes-crypt.class.js";
+export {SmpAesCryptV2} from "./smp-aes-crypt-v2.class.js";
 export {SmpCrypto} from "./smp-crypto.class.js";
 export {SmpSecureCode} from "./smp-secure-code.class.js";